+++ .env
@@ -1,0 +1,1 @@
+GEMINI_API_KEY=YOUR_API_KEY_HERE
